Switch to ShellCommand for checking out app-checker source.
It turns out that buildbot will override the default branch on source 
checkouts.  This makes it impossible to use the source management 
features in a multi-tree build, like app-checker, and have the 
build-on-change builds work properly.  So, we continue to use the 
ChangeSource mechanism to decide whether to build, but actually do the 
build by calling bzr manually.