Save old build RPMs instead of triggering new ones when needed.
In the past, we would trigger a new build every time one build depended
on another.  This looked difficult to port to the new buildbot, and since
there's a long-standing request to reuse old built packages instead,
we just do that now.

This commit now gets rid of the IndepTriggerable custom scheduler, which
we no longer need.